pyproject.toml
[project]
name = "agentic-rag"
version = "0.1.0"
description = "Agentic Retrieval for in RAG pipeline"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chonkie[hub]>=1.0.7",
    "joblib>=1.5.0",
    "langchain-core>=0.3.60",
    "langgraph>=0.4.5",
    "loguru>=0.7.3",
    "markitdown[all]>=0.1.1",
    "pandas>=2.2.3",
    "python-dotenv>=1.1.0",
    "tqdm>=4.67.1",
    "sentence-transformers>=3.3.1",
    "chromadb>=0.6.0",
    "rank-bm25>=0.2.2",
    "pdbpp>=0.10.3",
    "groq>=0.13.0",
    "langchain-groq>=0.2.0",
    "gradio>=5.0.0",
]

[tool.ruff]
line-length = 120
